Fix double value in pagination feature

+++ src/main/java/com/searchroom/service/PaginationService.java
@@ -2,7 +2,7 @@ package com.searchroom.service;
 
 public interface PaginationService {
 
-    double calculatePageAmount(int roomPerPage);
-    double calculatePageCustomerPost(int customerId, int roomPerPage);
+    int calculatePageAmount(int roomPerPage);
+    int calculatePageCustomerPost(int customerId, int roomPerPage);
 
 }
